LayoutTests/platform/mac/fast/images/support-broken-image-delegate-expected.txt

CONSOLE MESSAGE: TypeError: testRunner.setShouldPaintBrokenImage is not a function. (In 'testRunner.setShouldPaintBrokenImage(false)', 'testRunner.setShouldPaintBrokenImage' is undefined)
CONSOLE MESSAGE: TypeError: testRunner.setShouldPaintBrokenImage is not a function. (In 'testRunner.setShouldPaintBrokenImage(true)', 'testRunner.setShouldPaintBrokenImage' is undefined)
Radar 8610908 - Add a resource delegate method to query if WebCore should paint the broken image.

FAIL
support-broken-image-delegate.html - willSendRequest <NSURLRequest URL support-broken-image-delegate.html, main document URL support-broken-image-delegate.html, http method GET> redirectResponse (null)
support-broken-image-delegate.html - didReceiveResponse <NSURLResponse support-broken-image-delegate.html, http status code 0>
support-broken-image-delegate.html - didFinishLoading
resources/broken-image-with-invalid-format.png - willSendRequest <NSURLRequest URL resources/broken-image-with-invalid-format.png, main document URL support-broken-image-delegate.html, http method GET> redirectResponse (null)
resources/broken-image-with-invalid-format.png - didReceiveResponse <NSURLResponse resources/broken-image-with-invalid-format.png, http status code 0>
resources/broken-image-with-invalid-format.png - didFinishLoading
